Fix JvmName file usage
